refactor: align format of parsers with formatters
