+++ src/fields.d.ts
@@ -1,0 +1,7 @@
+export type RawField = {
+  [key: string]: string[];
+}
+
+export type Field = {
+  [key: string]: string;
+}
