+++ src/index.ts
@@ -1,7 +1,8 @@
 import {Command, flags} from '@oclif/command'
 
-import {parseSObjectFile, fieldProperties} from './parsers'
+import {getParser} from './parsers'
 import {getFormatter} from './formatters'
+import {RawField} from './fields'
 
 const FIELD_NAMES = [
   'label',
@@ -9,10 +10,6 @@ const FIELD_NAMES = [
   'type',
   'required',
 ]
-
-type RawField = {
-  [key: string]: string;
-}
 
 class ParseSalesforceObject extends Command {
   // TODO
@@ -56,16 +53,14 @@ class ParseSalesforceObject extends Command {
   async run() {
     const {args, flags} = this.parse(ParseSalesforceObject)
 
-    // parse sobject file
-    // const Parser = args.path.includes('object-meta.xml') ? SourceFormatParser : MetadataFormatParser
-    const rawFields = await parseSObjectFile(args.path)
+    const parser = getParser(args.path)
+    const rawFields = await parser.parse(args.path)
 
     const fields = this.extractFields(rawFields, FIELD_NAMES)
 
     const formatter = getFormatter(flags.format as 'soql' | 'markdown' |'csv')
-    // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-    // @ts-ignore
-    const result = formatter.format(fields, {namespace: flags.namespace, path: args.path, fieldNames: fieldProperties})
+
+    const result = formatter.format(fields, {namespace: flags.namespace, path: args.path, fieldNames: FIELD_NAMES})
     this.log(result)
   }
